Update ultimate.py
Add a gc.collect() call after file cleanup is done and before announcing that it is done.  I observed that it would fill up ram after several episodes so this is an attempt to prevent the scenario by having Python do a force garbage collection after each episode is done.